Major object-oriented redesign.
I created new direct.Simulation and expanded.Phase objects to hold
simulation data. The library functions for finding bulk coexistence
now work by accepting these objects and returning the solutions from
fsolve directly as the ratios of the new activities to the old. I also
added function to find, e.g., the composition of each of these
objects.

I moved some expanded ensemble-specific read_* functions to read.py.

The functions in wetting.py that find spreading/drying coefficients
now directly accept an interface potential, instead of reading data
from a specified directory.

The overall object of this redesign is to produce a leaner library
that forces the user to do a little bit more work in exchange for a
greater degree of flexibility.